% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/concordex.R
\name{calculateConcordex}
\alias{calculateConcordex}
\alias{calculateConcordex,ANY-method}
\title{Identify Spatially Homogeneous Regions with concordex}
\usage{
calculateConcordex(x, ...)

\S4method{calculateConcordex}{ANY}(
  x,
  labels,
  ...,
  n_neighbors = 30,
  compute_similarity = FALSE,
  BLUSPARAM,
  BNINDEX,
  BNPARAM = KmknnParam(),
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{A numeric matrix specifying the neighborhood structure of
observations. Typically an adjacency matrix produced by a k-Nearest
Neighbor algorithm. It can also be a matrix whose rows correspond to each
observation and columns correspond to neighbor indices, i.e. matrix form of
an adjacency list which can be a matrix due to fixed number of neighbors.}

\item{...}{Arguments passed to methods.}

\item{labels}{A numeric or character vector containing the label or class
corresponding to each observation. For example, a cell type or cluster ID.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing the metric for numerous observations shall be
parallelized.}

\item{k}{Number of neighbors to expect for each observation. Defaults to 20.}

\item{n.iter}{A number specifying the number of permutations for correcting
the coefficient.}

\item{return.map}{Logical, whether to return the matrix of the number of
cells of each label in the neighborhood of cells of each label.}
}
\value{
A named list with the following components:
\describe{
\item{\code{concordex}}{
The raw concordex coefficient corresponding to the original label assignments.
}

\item{\code{mean_random_concordex}}{
The average of \code{n.iter} concordex coefficients. concordex coefficients are computed
after permuting the labels and reassigning them to new observations.
}
\item{\code{corrected_concordex}}{
Simply the raw concordex coefficient divided by the average of the permuted
coefficients.
}
\item{\code{simulated}}{
Numeric vector of the concordex coefficients from permuted labels, showing the
null distribution.
}
\item{\code{map}}{
Numeric matrix of the number of cells of each label in the neighborhood of
cells of each label. Only returned when \code{return.map = TRUE}.
}
}
}
\description{
Compute the raw and corrected concordex coefficient using a
neighborhood graph and observation labels.
}
