% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nomap.R
\name{calculateNomap}
\alias{calculateNomap}
\alias{calculateNomap,ANY-method}
\title{Compute the Nomap coefficient}
\usage{
calculateNomap(x, ...)

\S4method{calculateNomap}{ANY}(
  x,
  labels,
  n.iter = 15,
  return.map = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{A numeric matrix specifying the neighborhood structure of observations.
Typically an adjacency matrix produced by a nearest-neighbor algorithm.}

\item{...}{Not currently used.}

\item{labels}{A numeric or character vector containing the label or class
corresponding to each observation. For example, a cell type or cluster ID.}

\item{n.iter}{A number specifying the number of permutations for correcting
the coefficient.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing the metric for numerous observations shall be parallelized.}
}
\value{
A named list with the following components:
\describe{
\item{\code{nomap}}{
The raw coefficient corresponding to the original label assignments.
}

\item{\code{mean_random_nomap}}{
The average of \code{n.iter} nomap coefficients. Nomap coefficients are computed
after after permuting the labels and reassigning them to new observations.
}
\item{\code{corrected_nomap}}{
Simply the raw nomap coefficient divided by the average of the permuted
coefficients.
}
}
}
\description{
Computes the raw and corrected Nomap coefficient using a neighborhood graph
and observation labels.
}
