% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{heatConcordex}
\alias{heatConcordex}
\title{Plot the concordex map matrix as a heatmap}
\usage{
heatConcordex(concordex, ...)
}
\arguments{
\item{concordex}{Output from \code{\link{calculateConcordex}}.}

\item{...}{Other arguments passed to \code{\link{pheatmap}} to customize the
plot.}
}
\value{
A \code{pheatmap} object.
}
\description{
The \code{calculateConcordex} function returns a matrix showing the number of
cells of each label in the neighborhood of cells of each label when argument
\code{return.map = TRUE}. This function plots this matrix as a heatmap, which
can be used as a clustering diagnostic.
}
\examples{
library(BiocNeighbors)
g <- findKNN(iris[, seq_len(4)], k = 10)
res <- calculateConcordex(g$index,
    labels = iris$Species, k = 10,
    return.map = TRUE
)
heatConcordex(res)

}
