% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{heatNomap}
\alias{heatNomap}
\title{Plot the nomap map matrix as a heatmap}
\usage{
heatNomap(nomap, ...)
}
\arguments{
\item{nomap}{Output from \code{\link{calculateNomap}}.}

\item{...}{Other arguments passed to \code{\link{pheatmap}} to customize the
plot.}
}
\value{
A \code{pheatmap} object.
}
\description{
The \code{calculateNomap} function returns a matrix showing the number of
cells of each label in the neighborhood of cells of each label when argument
\code{return.map = TRUE}. This function plots this matrix as a heatmap, which
can be used as a clustering diagnostic.
}
\examples{
library(BiocNeighbors)
g <- findKNN(iris[, seq_len(4)], k = 10)
res <- calculateNomap(g$index,
    labels = iris$Species, k = 10,
    return.map = TRUE
)
heatNomap(res)

}
