% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labels.R
\name{labels_guess_type}
\alias{labels_guess_type}
\title{Friendly types for continuous/discrete label combinations}
\usage{
labels_guess_type(labels)
}
\description{
Friendly types for continuous/discrete label combinations
}
