% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotNomapSim}
\alias{plotNomapSim}
\title{Plot density plot of simulated results}
\usage{
plotNomapSim(nomap, ...)
}
\arguments{
\item{nomap}{Output from \code{\link{calculateNomap}}.}

\item{...}{Other arguments passed to \code{\link{geom_density}}.}
}
\value{
A ggplot2 object. The density plot shows the simulated nomap
coefficient from permuted labels, while the vertical line shows the actual
nomap coefficient.
}
\description{
The nomap traces from permuted labels represent the null distribution of the
trace. This can be plotted as a density plot and visually compared to the
actual value.
}
\examples{
library(BiocNeighbors)
g <- findKNN(iris[, seq_len(4)], k = 10)
res <- calculateNomap(g$index, labels = iris$Species, k = 10)
plotNomapSim(res)

}
