% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_names_handler}
\alias{set_names_handler}
\title{Set names to margin of a matrix or update names of a vector}
\usage{
set_names_handler(allow_missing_nm = TRUE, allow_null_nm = TRUE, margin = 1L)
}
\description{
Set names to margin of a matrix or update names of a vector
}
