% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/concordex.R
\docType{methods}
\name{calculateConcordex}
\alias{calculateConcordex}
\alias{calculateConcordex,ANY-method}
\alias{calculateConcordex,SummarizedExperiment-method}
\alias{calculateConcordex,SingleCellExperiment-method}
\alias{calculateConcordex,SpatialExperiment-method}
\title{Identify Spatially Homogeneous Regions with concordex}
\usage{
calculateConcordex(x, ...)

\S4method{calculateConcordex}{ANY}(
  x,
  labels,
  ...,
  n_neighbors = 30,
  compute_similarity = FALSE,
  BLUSPARAM,
  BNINDEX,
  BNPARAM = KmknnParam(),
  BPPARAM = SerialParam()
)

\S4method{calculateConcordex}{SummarizedExperiment}(x, labels, ..., assay.type = "logcounts")

\S4method{calculateConcordex}{SingleCellExperiment}(x, labels, ..., use.dimred = NULL)

\S4method{calculateConcordex}{SpatialExperiment}(x, labels, ..., use.spatial = TRUE)
}
\arguments{
\item{x}{A \code{\link{SpatialExperiment}},
\code{\link{SpatialFeatureExperiment}},
\code{\link{SingleCellExperiment}}, or \code{\link{SummarizedExperiment}}
object containing a count matrix.

Otherwise, a numeric matrix-like object containing counts for cells
(or spots) in the columns and features in the rows.}

\item{...}{Other parameters passed to methods}

\item{labels}{Cell/spot labels used to compute the neighborhood consolidation
matrix. Continuous or discrete labels are allowed, and typically, integer
labels are assumed to be discrete.

Labels can be specified as follows:
\itemize{
\item A vector or matrix-like object with one entry per cell. If a matrix, the
cells should be on the rows and the label identifiers on the columns.
\item If \code{x} inherits from \code{\link{SummarizedExperiment-class}}, a
string or character vector specifying the names of columns in \code{colData(x)}
or the name of a dimensionality reduction result (see \link{reducedDimNames})
}}

\item{n_neighbors}{Number of neighbors to expect for each observation.
Defaults to 30.}

\item{compute_similarity}{Logical. Whether to return the label similarity matrix.
Only useful if discrete labels are provided.}

\item{BLUSPARAM}{A \code{\link{BlusterParam}} object specifying the clustering
algorithm to use to identify spatial homogeneous regions. If this parameter
is not specified, then regions are not returned. By default, this parameter
is missing.}

\item{BNINDEX}{A \code{\link{BiocNeighborIndex}} object containing the precomputed
index information.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing the metric for numerous observations shall be
parallelized (see \code{\link{bpparam}}).}

\item{assay.type}{String or integer scalar indicating the assay of \code{x}
containing the counts.}

\item{use.dimred}{Integer or string specifying the reduced dimensions to use
for construction of the k-nearest neighbor graph. Note that if this is not
\code{NULL}, reduced dimensions can not be used as labels to compute the
neighborhood consolidation matrix.}

\item{use.spatial}{Logical, should the spatial coordinates be used to compute the
k-nearest neighbor graph?}
}
\value{
A sparse matrix
}
\description{
Compute the neighborhood consolidation matrix and identify spatial
homogeneous regions.
}
