% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nomap.R
\name{calculateNomap}
\alias{calculateNomap}
\alias{calculateNomap,ANY-method}
\title{Compute the nomap coefficient}
\usage{
calculateNomap(x, ...)

\S4method{calculateNomap}{ANY}(
  x,
  labels,
  n.iter = 15,
  return.map = FALSE,
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{A numeric matrix specifying the neighborhood structure of observations.
Typically, an adjacency matrix produced by a nearest-neighbor algorithm.}

\item{...}{Not currently used.}

\item{labels}{A vector containing the label or class corresponding to each observation.
For example, a cell type or cluster id. Can be numeric or a character string.}

\item{n.iter}{Number of permutations.}

\item{BPPARAM}{A \code{\link{BiocParallelParam}} object specifying whether
and how computing the metric for numerous observations shall be parallelized.}
}
\description{
Compute the nomap coefficient
}
