% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labels.R
\name{labels_make_friendly}
\alias{labels_make_friendly}
\title{Collapse multiple discrete vectors and one-hot-encode}
\usage{
labels_make_friendly(labels, ..., sep = "_", nm = NULL)
}
\description{
Collapse multiple discrete vectors and one-hot-encode
}
